ceph_orch: run "ceph orch status" with 30-second timeout
The "ceph orch status" command is known to hang if /etc/ceph/ceph.conf
on the host points to hosts that used to be MONs but are not anymore.

It's debatable what an appropriate timeout is (30 seconds, more, less?)
but commands that are known to hang should have an explicit timeout.

References: https://github.com/ceph/ceph-salt/issues/199
Signed-off-by: Nathan Cutler <ncutler@suse.com>